requirements.txt
# Streamlit frontend
streamlit==1.30.0

# PDF/DOCX parsing
PyMuPDF==1.23.1
pdfplumber==0.9.0
python-docx==0.8.11
docx2txt==0.8

# NLP and embeddings
spacy==3.6.1
nltk==3.8.1
sentence-transformers==2.2.2
huggingface-hub==0.13.4
torch==2.2.0

# Vector database for semantic search
faiss-cpu==1.7.4

# Utilities
pandas==2.1.1
numpy==1.25.0
scikit-learn==1.3.1
tqdm==4.66.1

# Optional for advanced LLM feedback
openai==0.28.0
